Align tags and fix tests
